Corrected millis and micros to use uint64_t

+++ Acopter32-STM32F4/libraries/AP_InertialSensor/AP_InertialSensor_VRBRAIN.h
@@ -57,7 +57,7 @@ private:
 
     static const float          _gyro_scale;
 
-    uint32_t _last_sample_time_micros;
+    uint64_t _last_sample_time_micros;
 
     // ensure we can't initialise twice
     bool                        _initialised;
